fix(eslint-config): separate js and ts via extension
