--- index.js
@@ -1,5 +1,0 @@
-/* eslint-disable functional/immutable-data */
-
-module.exports = require("./react");
-
-/* eslint-enable functional/immutable-data */
